refactor(layout): one TileLayout core with per-platform hook objects instead of three near-identical copies
TileLayout.{win32,linux,darwin}.tsx were 597/535/520-line copies of the same
component; 437 of 598 lines were byte-identical across all three, and the
rest differed in a dozen known places. The shared body now lives once in
TileLayout.core.tsx as createTileLayout(platform); each platform file keeps
only its TileLayoutPlatform object and its own ResizeHandle:

  animateDelayMs               win32 150 (Issue #774), linux/darwin 50
  onRootMount                  win32: the Win11 dragend safety net
  boundsCheckRequiresTileDrag  linux: skip the dragover bounds check outside
                               a tile drag (tab DnD crashed persistToBackend)
  onLeafGeometryChange         win32: notifyPaneReflow for native panes
  clearDraggingWhenInactive    win32: drop .dragging after the safety net
  rejectDragAt                 win32: no tear-off from a resize-handle zone
  before/afterDragStart,       linux/darwin: preventUnhandled start/stop;
  before/afterDrop             linux: setJsDragActive(true/false)
  ResizeHandle                 pointer capture (win32), capture + body-cursor
                               workaround (darwin), window listeners (linux)

Every hook sits exactly where the per-platform code was, so each build's
behaviour is unchanged; the drag REGISTRATION (draggable() on the live header
found by querySelector, 100 ms re-register poll, never torn down mid-drag)
was already the same on all three engines and moves to the core as is.

1,652 -> 957 lines. tilelayout-shared.tsx's header no longer claims
DisplayNode is per-platform.
