+++ .changesets/1788751173-tilelayout-one-shared-core-createtilelayout-with-per-platfor-jnel.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+TileLayout: one shared core (createTileLayout) with per-platform hook objects for win32/linux/darwin instead of three near-identical copies; no behavior change
